--- _pytest/helpers/__init__.py
@@ -1,5 +1,0 @@
-class ResponseTest(object):
-    def __init__(self, endpoint, expected_response, payload=None):
-        self.endpoint = endpoint
-        self.expected_response = expected_response
-        self.payload = payload
