+++ src/tokenizers/whitespace_tokenizer.py
@@ -2,5 +2,5 @@ from rasa_nlu.tokenizers import Tokenizer
 
 
 class WhitespaceTokenizer(Tokenizer):
-    def tokenize(self, text, nlp=None):
+    def tokenize(self, text):
         return text.split()
